Update prepare_sector_network.py
biochar heat can be wested (directed to a store).
before the store was connected on the biochar_heat_bus, hence connected to the central urban heat. 
now the store is connected to a new bus "biochar_heat_out", and a new link from the buses "biochar_heat_bus" and "biochar_heat_out" is created. 
In this way the heat accumulated in the store cannot be later sent to the district heating